ml_sample/investment-meeting/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "investment-meeting"
version = "0.1.0"
description = "Multi-agent investment decision simulator with Gemini 2.5 Flash Lite"
requires-python = ">=3.10"
authors = [{name = "Investment AI Lab"}]

dependencies = [
    "google-genai>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "asyncio-contextmanager>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
lint.select = ["E", "F", "W", "I", "UP", "B", "C4"]
lint.ignore = ["E501", "W605"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
